fix many warnings

+++ libomptarget/plugins/cloud/src/providers/generic.h
@@ -18,7 +18,7 @@
 
 struct HdfsInfo {
   std::string ServAddress;
-  int ServPort;
+  uint16_t ServPort;
   std::string UserName;
 };
